Properly truncate incoming messages from rpi
The input string is now truncated each time a full message has
arrived; previous behavior was the gradual increase in size of
`inputString`, until it was cleared, which might lead to truncation
of long messages and dropping of commands.

This change has been tested on all arduinos except the `PD` one. The
`STIR` arduino still seems to drop a command every ~10, so it
might need a few more tests to remove all possible corner cases, which
may be due to the `Serial1` messages overlapping.