Update evolver_si to properly parse last instance of serial input
Stir commands were being missed due to the code not being properly parsed.

quick summary: The string wasn’t interpreted correctly if there were additional commands being appended on it. For example, if somehow the timing got into a certain loop, the string would end up like this

stire,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,endstira,,,,,,,,,,,,,,,,,_!tempr,1878,1882,1875,1888,1885,1879,1878,1869,1868,1880,1862,1872,1884,1864,1879,1881,_!
the ack would never be run because it would interpret the last “stir” and the last _!
meaning this whole chunk would be tried to be interpreted (incorrectly):
stira,,,,,,,,,,,,,,,,,_!tempr,1878,1882,1875,1888,1885,1879,1878,1869,1868,1880,1862,1872,1884,1864,1879,1881,_!
this doesnt happen often but happens for stir for some reason
I changed the library to interpret it correctly, looking for the closest _! after the last  address
so everyone needs up update their libraries